+++ NEWS.md
@@ -1,3 +1,5 @@
+# ipaddress (development version)
+
 # ipaddress 0.5.1
 
 * Addition/subtraction of large integers is now much faster (#64)