+++ NEWS.md
@@ -5,7 +5,7 @@
   * `base = "bin"` is now handled by the existing `ip_to_binary()` function.
   * `base = "hex"` is now handled by the new `ip_to_hex()` function.
 
-* `ip_to_integer()` and `integer_to_ip()` now use `bignum::biginteger()` vectors to store IP addresses as integers. Previously, these integers were stored in a character vector (because they were beyond the range covered by base R numeric types).
+* `ip_to_integer()` and `integer_to_ip()` now use `bignum::biginteger()` vectors to store IP addresses as integers. Previously, these integers were stored in a character vector (because they were beyond the range of base R numeric types).
   * The BH package is no longer a dependency (C++ Boost headers were a heavy dependency).
   * The bignum package is now an optional dependency.