+++ NEWS.md
@@ -2,7 +2,7 @@
 
 * The `base` argument of `ip_to_integer()` is removed.
   * `base = "dec"` is still handled by `ip_to_integer()`.
-  * `base = "bin"` is now handled by the existing `ip_to_binary()` function.
+  * `base = "bin"` is still handled by the existing `ip_to_binary()` function.
   * `base = "hex"` is now handled by the new `ip_to_hex()` function.
 
 * `ip_to_integer()` and `integer_to_ip()` now use `bignum::biginteger()` vectors to store IP addresses as integers. Previously, these integers were stored in a character vector (because they were beyond the range of base R numeric types).