+++ NEWS.md
@@ -1,4 +1,4 @@
-# ipaddress (development version)
+# ipaddress 0.5.2
 
 * The `base` argument of `ip_to_integer()` is removed.
   * `base = "dec"` is still handled by `ip_to_integer()`.